TESTS: Split scan into scan and scan by key files
